backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5
openai>=1.0.0
numpy>=1.21.0
tenacity>=8.0.0
GitPython>=3.1.0
aiohttp>=3.8.0
